# frontend/openapi.yaml
openapi: 3.0.3
info:
  title: Medclara Backend API
  description: |
    # Medclara Clinical Documentation Platform API
    
    Complete API specification for the Medclara backend, including:
    - **Authentication** - User login and session management
    - **Patients** - Patient data management
    - **Templates** - Clinical note templates
    - **Sessions** - Scribe session management
    - **Recordings** - Audio recording upload and processing
    - **Chunked Upload** - Large file upload with resumable chunks
    - **Server-Sent Events (SSE)** - Real-time processing status updates
    - **Notes** - Clinical note creation and management
    - **Workspace** - Scribe workspace API
    
    ## Real-Time Processing with SSE
    
    The Medclara API uses Server-Sent Events (SSE) for real-time recording processing updates.
    This provides 99% fewer network requests and 20x faster response times compared to polling.
    
    ### Complete Recording Workflow
    1. **POST** `/api/recordings/chunked/init` - Initialize upload session
    2. **POST** `/api/recordings/chunked/upload` - Upload audio chunks (repeat)
    3. **POST** `/api/recordings/chunked/finalize` - Start processing
    4. **GET** `/api/recordings/{recordingId}/stream` - Stream processing events (SSE)
    5. Process events: `processing` → `completed` or `failed`
    
    ### Benefits of SSE Over Polling
    - **99% fewer requests** - One persistent connection vs. polling every 500ms
...
    
  version: '1.0.0'
...
servers:
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from login endpoint
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          description: Error code for programmatic handling
          examples:
            - unauthorized
            - not_found
            - invalid_request
            - server_error
        message:
          type: string
          description: Human-readable error message
        details:
          type: object
          description: Additional error details
          additionalProperties: true
...
    Template:
      type: object
      required:
        - id
        - template_key
        - label
        - sections
        - is_active
      properties:
        id:
          type: string
          format: uuid
          description: Unique template identifier
        template_key:
          type: string
          description: Machine-readable template key (e.g., 'soap-general')
          pattern: '^[a-z0-9\-]+$'
        label:
          type: string
          description: Human-readable template name
        description:
          type: string
        specialty:
          type: string
          description: Medical specialty (e.g., 'cardiology', 'neurology')
        category:
          type: string
          description: Template category (e.g., 'General', 'Specialized Consultations')
        sections:
          type: array
          description: Template sections for clinical notes
          items:
            type: object
            required:
              - key
              - title
            properties:
              key:
                type: string
              title:
                type: string
              helper:
                type: string
              order:
                type: integer
        is_active:
          type: boolean
          description: Whether template is available for use
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        # Prompt-based template fields
        prompt:
          type: string
          description: AI prompt for extraction
        extract_style:
          type: string
          enum: [narrative, structured, hybrid]
        prompt_version:
          type: integer
...
paths:
...
  /api/templates/key/{templateKey}:
    get:
      tags:
        - Templates
      summary: Get template by key
      description: Retrieve template by machine-readable key (e.g., 'soap-general')
      operationId: getTemplateByKey
      security:
        - BearerAuth: []
      parameters:
        - name: templateKey
          in: path
          required: true
          schema:
            type: string
            description: Template key (e.g., 'soap-general', 'soap-cardiology')
      responses:
        '200':
          description: Template details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'
        '404':
          description: Template not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
security:
  - BearerAuth: []